requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.10.0
babel==2.17.0
beautifulsoup4==4.14.2
cachetools==6.2.4
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
courlan==1.3.2
cssselect==1.3.0
dateparser==1.2.2
distro==1.9.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
jieba3k==0.35.1
joblib==1.5.2
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.3.5
pandas==2.3.3
pillow==12.0.0
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytrends==4.9.2
pytz==2025.2
PyYAML==6.0.3
readability-lxml==0.8.4.1
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
rsa==4.9.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
tenacity==9.1.2
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.0
tqdm==4.67.1
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
qdrant-client==1.13.2
